feat: Extract account generation to samples_common
