+++ sign/samples_common/build.gradle.kts
@@ -48,6 +48,7 @@ dependencies {
 
     lifecycle()
     navigationComponent()
+    api(project(":androidCore:sdk"))
 
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
