+++ sign/samples_common/src/main/kotlin/com/walletconnect/sample_common/Keys.kt
@@ -1,4 +1,4 @@
-package com.walletconnect.web3.inbox.domain
+package com.walletconnect.sample_common
 
 import com.walletconnect.foundation.common.model.PrivateKey
 import com.walletconnect.foundation.common.model.PublicKey
@@ -7,9 +7,10 @@ import org.bouncycastle.jce.provider.BouncyCastleProvider
 import org.web3j.crypto.Keys
 import java.security.Security
 
-internal fun generateEthereumAccount(address: String) = "eip155:1:0x$address"
 
-internal fun generateKeys(): Triple<String, String, String> {
+fun generateEthereumAccount(address: String) = "eip155:1:0x$address"
+
+fun generateKeys(): Triple<String, String, String> {
     Security.removeProvider("BC")
     Security.addProvider(BouncyCastleProvider())
     val keypair = Keys.createEcKeyPair()
@@ -17,7 +18,3 @@ internal fun generateKeys(): Triple<String, String, String> {
     val privateKey = PrivateKey(keypair.privateKey.toByteArray().bytesToHex())
     return Triple(publicKey.keyAsHex, privateKey.keyAsHex, Keys.getAddress(keypair))
 }
-
-internal const val ACCOUNT_TAG = "self_account_tag"
-internal const val PRIVATE_KEY_TAG = "self_private_key"
-internal const val PUBLIC_KEY_TAG = "self_public_key"
