+++ web3/inbox/src/main/kotlin/com/walletconnect/web3/inbox/proxy/request/GetReceivedInvitesRequestUseCase.kt
@@ -23,6 +23,4 @@ internal class GetReceivedInvitesRequestUseCase(
         Web3InboxParams.Response.GetReceivedInvitesResult(id, inviterAccount.value, inviteeAccount.value, message.value, inviterPublicKey, inviteePublicKey, status.toResult())
 
     private fun Chat.Type.InviteStatus.toResult(): String = name.lowercase()
-
 }
-
